Recursive lookup in envs when resolving
uncaught tokens was not working! Problem was:
In parse My starting point env was not the real-actual toplevel env, but
a DEEPCOPY thereof! Appending to that copy as enclosingenv in parse
resulted of course in no change in the toplevel env. Using not a copy
but the real toplevle env solves this!