fixed closures problem:
When evaluating, we eval always in each block's own env (and never in
TOPLEVEL env, which was the mistake). Since envs always have
child-parent relations which resolve_token method goes through (all the
way up to the TLENV), there is no need to directly eval in tlenv.
The different btwn tlenv and lexical envs only matters by the question:
in which env is a name to be saved!
This gives a powerful closuring possibility to nest tl namings in
lexical namings eg.:

name
  tmp1 10
  tmp2 name
        tl ja
        var * tmp1 20

which writes only VAR as toplevel, and keeps tmp1 tmp2 as lexical (they
are thrown away!)