bugfix: set current_binding_unit to NULL
instead of finding previous binding unit. this forces
current binding unit to be recomputed if neccessary for next
unit which is safer in a context like the following:

let :x 30
  let :y x
    let :q lambda
             let :w y
               let :c w
                 lambda let :r c
                          lambda pret r
       define F q

here the let in line 4 must close it's lambda (line 3). setting
current_binding_unit back to the let on line 3 (the effect of
the find_prev_binding_unit function) :w would go into the let/3
which is trash.