bugfix #5: stop using intermediate Arv_data pointers in evaluations
in many eval_... functions there is an intermediate struct Arv_data*
which saves the final evaluation, to then assign to the actual struct
Arv_data *return_data. this is almost always unnecessary, since the
evaluation can directly be assigned to the return_data.  the eval3
function also had similiar issues when looking up names, which could
be fixed by assigning directly to it's return_data.
NOTE: this bug is not fixed for ALL evaluators, but only fixed here
for eval_define, eval3, eval_show_op!